add a base client for reusable code, rename all current clients/services to include claims
