+++ modules/travel_pay/app/controllers/travel_pay/v0/claims_controller.rb
@@ -16,7 +16,7 @@ module TravelPay
       private
 
       def service
-        @service ||= TravelPay::Service.new
+        @service ||= TravelPay::ClaimsService.new
       end
 
       def common_exception(e)
