+++ modules/travel_pay/spec/services/claims_service_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe TravelPay::Service do
+describe TravelPay::ClaimsService do
   context 'get_claims' do
     let(:user) { build(:user) }
     let(:claims_data) do
@@ -55,7 +55,7 @@ describe TravelPay::Service do
     end
 
     before do
-      allow_any_instance_of(TravelPay::Client)
+      allow_any_instance_of(TravelPay::ClaimsClient)
         .to receive(:get_claims)
         .with(user)
         .and_return(claims_response)
@@ -64,7 +64,7 @@ describe TravelPay::Service do
     it 'returns sorted and parsed claims' do
       expected_statuses = ['In Progress', 'In Progress', 'Incomplete', 'Claim Submitted']
 
-      service = TravelPay::Service.new
+      service = TravelPay::ClaimsService.new
       claims = service.get_claims(user)
       actual_statuses = claims[:data].pluck(:claimStatus)
